+++ examples/interact/test_program.ml
@@ -4,8 +4,7 @@ let some_value = [A 42; B "Hello, world"]
 
 let () =
   print_endline "Starting utop now!";
-  UTop_main.interact
-    ~search_path:["_build"]
+  UTop_main.interact ()
     ~unit:__MODULE__
     ~loc:__POS__
     ~values:[V ("some_value", some_value)]